This fixes a bug where colonies couldn't be placed anymore
And also fixes in 1.10 the bug where the message is called 4 times.

+++ src/main/java/com/minecolonies/colony/Colony.java
@@ -368,7 +368,7 @@ public class Colony implements IColony
     }
 
     @Override
-    public float getDistanceSquared(@NotNull BlockPos pos)
+    public long getDistanceSquared(@NotNull BlockPos pos)
     {
         //  Perform a 2D distance calculation, so pass center.posY as the Y
         return BlockPosUtil.getDistanceSquared(center, new BlockPos(pos.getX(), center.getY(), pos.getZ()));
